+++ src/main/java/racingcar/application/RacingService.java
@@ -1,4 +1,4 @@
-package racingcar.service;
+package racingcar.application;
 
 import java.util.List;
