+++ src/main/java/racingcar/ui/OutputView.java
@@ -1,4 +1,4 @@
-package racingcar.view;
+package racingcar.ui;
 
 import java.util.List;
